feat: Add custom build and run arguments for Go server
This change adds the ability to customize the arguments used for building and running the Go server. Previously, the server used default values for the build and run commands, but now users can specify their own custom arguments.

The new `CompileArgs` and `RunArgs` fields allow users to pass in a list of arguments to be used with the `go build` and `go run` commands, respectively. This gives more flexibility in how the server is built and executed.

Additionally, the `getBinaryName()` and `getGoBuildCMD()` methods have been added to encapsulate the logic for determining the binary name and constructing the build command.