fix: Update Dagger versions in tests
This commit updates the Dagger versions used in the test suite. Specifically, the `TestDaggerWithDaggerCLI` function now tests the latest Dagger versions from `v0.13.0` to `v0.13.3`, and the `TestDaggerSetupDaggerInDagger` function uses Dagger `v0.13.3`.

Additionally, the `TestAll` function in `main.go` has been updated to use a maximum of 10 goroutines and to cancel the test pool on the first error. The commented-out code for handling errors has also been updated.

The `golang.go` file has been updated to remove the early return for an empty list of packages in the `WithGoInstall` function.