feat: Add support for mounting configuration files in the container
This change introduces a new method `WithConfigFile` to the `ModuleTemplate` struct, which allows you to mount a configuration file into the container at a specified path. If no path is provided, it defaults to a predefined mount prefix from the `fixtures` package.

The method also supports setting an environment variable in the container with the path of the mounted configuration file.

This new functionality will simplify the process of managing and accessing configuration files within the container, improving the overall developer experience.