feat: Add module type option to create command
The changes in this commit add a new `module_type` option to the `create` command in the `main.rs` file. The `module_type` parameter allows the user to specify the type of module to be created, with a default value of "full" if not provided. This gives the user more flexibility in creating different types of modules using the `daggy` tool.