feat: Add new recipe to create modules with optional CI setup
Introduce a new recipe `create` that allows creating a new module with Daggy.
The recipe now has an optional `with-ci` parameter that, when set to `true`,
will also run the `cilocal` recipe for the newly created module.
Additionally, the `createlight` recipe has also been updated to support the
`with-ci` parameter.