fix: Add a new recipe to list functions in examples/go module
The changes in this commit add a new recipe called `listfnsexamples` to the Justfile. This recipe is used to list the available functions in the `examples/go` module of a given module. 

The new recipe follows the same structure as the existing `listfns` and `listfnstest` recipes, but it changes the directory to the `examples/go` folder instead of the `tests` folder. This allows users to conveniently retrieve the list of functions in the examples module.